Split code up into modules
