+++ pytest_sentry/__init__.py
@@ -1,0 +1,9 @@
+from .client import Client
+from .fixtures import sentry_test_scope
+from .pytest_hooks import (
+    pytest_fixture_setup,
+    pytest_load_initial_conftests,
+    pytest_runtest_call,
+    pytest_runtest_makereport,
+    pytest_runtest_protocol,
+)
